pyproject.toml
[tool.poetry]
name = "less-7"
version = "0.1.0"
description = ""
authors = ["Aidar Putilov <putiloff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
black = "^24.10.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.7"
drf-yasg = "^1.21.8"
stripe = "^11.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
